Made the job details slider its own component Added scroll and maximize button to the job details slider
