+++ package.json
@@ -20,6 +20,11 @@
   },
   "homepage": "https://github.com/singnet/snet-dapp#readme",
   "dependencies": {
+    "@fortawesome/fontawesome-svg-core": "^1.2.12",
+    "@fortawesome/free-brands-svg-icons": "^5.6.3",
+    "@fortawesome/free-regular-svg-icons": "^5.6.3",
+    "@fortawesome/free-solid-svg-icons": "^5.6.3",
+    "@fortawesome/react-fontawesome": "^0.1.3",
     "abi-decoder": "^1.1.0",
     "bitcoinjs-lib": "^3.3.2",
     "file-loader": "^1.1.11",
@@ -33,6 +38,7 @@
     "react-bootstrap": "^0.32.1",
     "react-copy-to-clipboard": "^5.0.1",
     "react-dom": "^16.3.1",
+    "react-perfect-scrollbar": "^1.4.4",
     "styled-components": "^3.2.5",
     "underscore": "^1.9.1"
   },
