+++ src/App.js
@@ -1,8 +1,13 @@
 import React from 'react';
 import {Link} from 'react-router-dom';
+import { library } from '@fortawesome/fontawesome-svg-core'
+import { fab } from '@fortawesome/free-brands-svg-icons'
+import { faCheckSquare, faCoffee } from '@fortawesome/free-solid-svg-icons'
+
 export default class App extends React.Component {
   constructor(props) {
-    super(props); 
+    super(props);
+    library.add(fab, faCheckSquare, faCoffee);
   }
 
   render() {
